api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Nicolas Landier <nicolas.landier@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
Jinja2 = "^2.11.2"
uvicorn = {extras = ["standard"], version = "^0.13.2"}
influxdb = "^5.3.1"
python-binance = "^0.7.9"
binance-sdk = "^1.2.7"
databases = {extras = ["sqlite"], version = "^0.4.3"}
SQLAlchemy = "1.3.23"
icecream = "^2.1.0"
pandas = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
